move alternate type test cases into a separate file
